move app functions out
