+++ app.js
@@ -1,45 +1,14 @@
 const express = require("express");
-const path = require("path");
 const indexRouter = require("./routers/indexRouter.js");
 const blogRouter = require("./routers/blogRouter.js");
 const sequelize = require("./config/database.js");
+const setupViewEngine = require("./setup/setupViewEngine.js");
+const setupMiddleware = require("./setup/setupMiddleware.js");
+const setupRoutes = require("./setup/setupRoutes.js");
+const syncDatabase = require("./setup/syncDatabase.js");
+const startServer = require("./setup/startServer.js");
 
-const app = express();
-const port = process.env.PORT || 3000;
-
-function setupViewEngine(app) {
-  app.set("views", path.join(__dirname, "views"));
-  app.set("view engine", "pug");
-}
-
-function setupMiddleware(app) {
-  app.use(express.urlencoded({ extended: true }));
-  app.use(express.static(path.join(__dirname, "public")));
-}
-
-function setupRoutes(app, routers) {
-  for (const router of routers) {
-    app.use(router.basePath, router.router);
-  }
-}
-
-function startServer() {
-  app.listen(port, () => {
-    console.log(`Server is running on http://localhost:${port}`);
-  });
-}
-
-async function syncDatabase() {
-  try {
-    await sequelize.sync();
-    console.log("Database synchronized");
-  } catch (error) {
-    console.error("Failed to sync database:", error);
-    process.exit(1);
-  }
-}
-
-async function initializeServer() {
+async function initializeServer(app, port) {
   setupViewEngine(app);
   setupMiddleware(app);
   setupRoutes(app, [
@@ -47,8 +16,11 @@ async function initializeServer() {
     { basePath: "/blog", router: blogRouter },
   ]);
 
-  await syncDatabase();
-  startServer();
+  await syncDatabase(sequelize);
+  startServer(app, port);
 }
 
-initializeServer();
+const app = express();
+const port = process.env.PORT || 3000;
+
+initializeServer(app, port);
