+++ setup/setupRoutes.js
@@ -1,0 +1,7 @@
+function setupRoutes(app, routers) {
+  for (const router of routers) {
+    app.use(router.basePath, router.router);
+  }
+}
+
+module.exports = setupRoutes;
