+++ setup/startServer.js
@@ -1,0 +1,7 @@
+function startServer(app, port) {
+  app.listen(port, () => {
+    console.log(`Server is running on http://localhost:${port}`);
+  });
+}
+
+module.exports = startServer;
